requirements.txt
openai>=1.0.0
streamlit>=1.30.0
transformers>=4.36.0
torch>=2.0.0
datasets>=2.14.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
accelerate>=0.24.0
